Update compatibility to numpy - 1.21.5 (#1033)
* Issue #670: Updates for numpy 1.20.0 compatibility

* Issue #670: Updating setup.py to allow up to numpy 1.21.5

+++ tests/join_test.py
@@ -71,7 +71,7 @@ class JoinTest(ArkoudaTest):
         self.assertEqual(0, I.size)
         self.assertEqual(0, J.size)
         
-        I,J = ak.join_on_eq_with_dt(self.a2,self.a1,self.t1,self.t2,dt,"pos_dt", np.int(0))
+        I,J = ak.join_on_eq_with_dt(self.a2,self.a1,self.t1,self.t2,dt,"pos_dt", int(0))
         self.assertEqual(0, I.size)
         self.assertEqual(0, J.size)
